# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для отслеживания стоимости криптовалют.",
        "title": "CryptoWatcher API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/currency/add": {
            "post": {
                "description": "Добавляет новую валюту для отслеживания.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Добавить валюту",
                "parameters": [
                    {
                        "description": "Данные для добавления валюты",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CurrencyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ\" example({\"message\": \"Currency added\"})",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/currency/price": {
            "get": {
                "description": "Возвращает цену валюты на указанный timestamp.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Получить цену валюты",
                "parameters": [
                    {
                        "description": "Данные для получения цены",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetPriceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ\" example({\"price\": 97675})",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "number"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/currency/remove": {
            "post": {
                "description": "Удаляет валюту из списка отслеживаемых.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Удалить валюту",
                "parameters": [
                    {
                        "description": "Данные для удаления валюты",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CurrencyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ\" example({\"message\": \"Currency removed\"})",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CurrencyRequest": {
            "type": "object",
            "properties": {
                "coin": {
                    "description": "Название валюты.",
                    "type": "string"
                }
            }
        },
        "model.GetPriceRequest": {
            "type": "object",
            "properties": {
                "coin": {
                    "description": "Название валюты.",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Временная метка (timestamp) записи.",
                    "type": "integer"
                }
            }
        }
    }
}